requirements.txt
matplotlib==3.5.1
numpy==1.21.6
pandas==1.4.3
Pillow==9.2.0
pmdarima==2.0.1
pycaret==3.0.0rc3
requests==2.28.1
scikit_learn==1.1.2
statsmodels==0.13.2
streamlit==1.11.1
streamlit_lottie==0.0.3
